+++ packages/polaris-core/src/server/polaris-server.ts
@@ -110,6 +110,8 @@ export class PolarisServer {
         const schema: GraphQLSchema = makeExecutablePolarisSchema(
             this.polarisServerConfig.enableFederation,
             this.polarisServerConfig.typeDefs,
+            this.polarisServerConfig.polarisSchemaConfig,
+            this.polarisServerConfig.permissionsConfig !== undefined,
             this.polarisServerConfig.resolvers,
             this.polarisServerConfig.schemaDirectives,
         );