+++ packages/polaris-schema/src/config/polaris-schema-config.ts
@@ -1,0 +1,9 @@
+export interface PolarisSchemaConfig {
+    addPolarisGraphQLScalars?: boolean;
+    polarisTypeDefs?: PolarisTypeDefsConfiguration;
+}
+
+export interface PolarisTypeDefsConfiguration {
+    addOnlinePagingTypeDefs?: boolean;
+    addFiltersTypeDefs?: boolean;
+}